refactor:revert moving httpError from types.ts to index.ts
