+++ lib/datasource/docker/index.ts
@@ -10,8 +10,8 @@ import type { HostRule } from '../../types';
 import { ExternalHostError } from '../../types/errors/external-host-error';
 import * as packageCache from '../../util/cache/package';
 import * as hostRules from '../../util/host-rules';
-import { Http, HttpError } from '../../util/http';
-import type { HttpOptions, HttpResponse } from '../../util/http/types';
+import type { Http } from '../../util/http';
+import { HttpError, HttpOptions, HttpResponse } from '../../util/http/types';
 import type { OutgoingHttpHeaders } from '../../util/http/types';
 import { hasKey } from '../../util/object';
 import { regEx } from '../../util/regex';
