+++ lib/util/http/index.ts
@@ -1,7 +1,6 @@
 import crypto from 'crypto';
 import merge from 'deepmerge';
 import got, { Options, Response } from 'got';
-import { RequestError as RequestError_ } from 'got';
 import { HOST_DISABLED } from '../../constants/error-messages';
 import { pkg } from '../../expose.cjs';
 import { logger } from '../../logger';
@@ -24,8 +23,6 @@ import type {
 } from './types';
 // TODO: refactor code to remove this (#9651)
 import './legacy';
-
-export { RequestError_ as HttpError };
 
 function cloneResponse<T extends Buffer | string | any>(
   response: HttpResponse<T>
